requirements.txt
aiogram==3.3.0
flask==2.3.2
python-dotenv==1.0.0